# src/Http.API/ClientApp/src/app/share/models/entity-library/entity-library-filter-dto.model.ts
/**
 * 实体库查询筛选
 */
export interface EntityLibraryFilterDto {
  pageIndex: number;
  pageSize: number;
  /**
   * 排序
   */
  orderBy?:  | null;
  name?: string | null;
  /**
   * 是否公开
   */
  isPublic?: boolean | null;

}