# src/Http.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "dusi.dev",
    "description": "API 文档",
    "version": "v1"
  },
  "paths": {
...
    "/api/EntityLibrary/filter": {
      "post": {
        "tags": [
          "EntityLibrary"
        ],
        "operationId": "EntityLibrary_Filter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityLibraryFilterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityLibraryFilterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EntityLibraryFilterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibraryItemDtoPageList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibraryItemDtoPageList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibraryItemDtoPageList"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/EntityLibrary/{id}": {
      "put": {
        "tags": [
          "EntityLibrary"
        ],
        "operationId": "EntityLibrary_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityLibraryUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityLibraryUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EntityLibraryUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibrary"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibrary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibrary"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "EntityLibrary"
        ],
        "operationId": "EntityLibrary_GetDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibrary"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibrary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibrary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EntityLibrary"
        ],
        "operationId": "EntityLibrary_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibrary"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibrary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityLibrary"
                }
              }
            }
          }
        }
      }
    },
    "/api/EntityMember/filter": {
      "post": {
        "tags": [
          "EntityMember"
        ],
        "operationId": "EntityMember_Filter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberFilterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberFilterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberFilterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberItemDtoPageList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberItemDtoPageList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberItemDtoPageList"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/EntityMember/{id}": {
      "put": {
        "tags": [
          "EntityMember"
        ],
        "operationId": "EntityMember_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMember"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMember"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMember"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "EntityMember"
        ],
        "operationId": "EntityMember_GetDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMember"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMember"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMember"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EntityMember"
        ],
        "operationId": "EntityMember_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMember"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMember"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMember"
                }
              }
            }
          }
        }
      }
    },
    "/api/EntityMemberConstraint/filter": {
      "post": {
        "tags": [
          "EntityMemberConstraint"
        ],
        "operationId": "EntityMemberConstraint_Filter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberConstraintFilterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberConstraintFilterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberConstraintFilterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraintItemDtoPageList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraintItemDtoPageList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraintItemDtoPageList"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/EntityMemberConstraint/{id}": {
      "put": {
        "tags": [
          "EntityMemberConstraint"
        ],
        "operationId": "EntityMemberConstraint_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberConstraintUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberConstraintUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EntityMemberConstraintUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraint"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraint"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraint"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "EntityMemberConstraint"
        ],
        "operationId": "EntityMemberConstraint_GetDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraint"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraint"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraint"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EntityMemberConstraint"
        ],
        "operationId": "EntityMemberConstraint_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraint"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraint"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityMemberConstraint"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessModifier": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "description": "访问修饰符",
        "format": "int32",
        "x-enumNames": [
          "Public",
          "Protected",
          "Internal"
        ]
      },
...
      "EntityLibrary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "软删除"
          },
          "name": {
            "maxLength": 60,
            "type": "string",
            "description": "库名称",
            "nullable": true
          },
          "description": {
            "maxLength": 200,
            "type": "string",
            "description": "库描述",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean",
            "description": "是否公开"
          },
          "entityModels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityModel"
            },
            "description": "包含的实体类",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/SystemUser"
          }
        },
        "additionalProperties": false,
        "description": "实体库"
      },
...
      "EntityLibraryFilterDto": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "orderBy": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean",
              "nullable": true
            },
            "description": "排序",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean",
            "description": "是否公开",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "实体库查询筛选"
      },
...
      "EntityLibraryItemDtoPageList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityLibraryItemDto"
            },
            "nullable": true
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "EntityLibraryUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 200,
            "type": "string",
            "description": "库描述",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean",
            "description": "是否公开",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "实体库更新时请求结构"
      },
      "EntityMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "软删除"
          },
          "name": {
            "maxLength": 60,
            "type": "string",
            "description": "属性名",
            "nullable": true
          },
          "comment": {
            "maxLength": 300,
            "type": "string",
            "description": "属性注释内容",
            "nullable": true
          },
          "defaultValue": {
            "maxLength": 100,
            "type": "string",
            "description": "默认值，根据类型推断",
            "nullable": true
          },
          "accessModifier": {
            "$ref": "#/components/schemas/AccessModifier"
          },
          "isRequired": {
            "type": "boolean",
            "description": "是否必须"
          },
          "isEnum": {
            "type": "boolean",
            "description": "是否为枚举"
          },
          "isList": {
            "type": "boolean",
            "description": "是否为列表"
          },
          "isObject": {
            "type": "boolean",
            "description": "是否为自定义对象"
          },
          "canSet": {
            "type": "boolean",
            "description": "是否可赋值"
          },
          "needInit": {
            "type": "boolean",
            "description": "是否要初始化"
          },
          "dictionaryKeyType": {
            "$ref": "#/components/schemas/MemberType"
          },
          "dictionaryValueType": {
            "$ref": "#/components/schemas/MemberType"
          },
          "memberType": {
            "$ref": "#/components/schemas/MemberType"
          },
          "constraint": {
            "$ref": "#/components/schemas/EntityMemberConstraint"
          },
          "objectType": {
            "$ref": "#/components/schemas/EntityModel"
          },
          "objectTypeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "entityModel": {
            "$ref": "#/components/schemas/EntityModel"
          }
        },
        "additionalProperties": false,
        "description": "实体属性"
      },
...
      "EntityMemberConstraintFilterDto": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "orderBy": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean",
              "nullable": true
            },
            "description": "排序",
            "nullable": true
          },
          "entityMemberId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "属性的约束查询筛选"
      },
...
      "EntityMemberConstraintItemDtoPageList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityMemberConstraintItemDto"
            },
            "nullable": true
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "EntityMemberConstraintUpdateDto": {
        "type": "object",
        "properties": {
          "maxLength": {
            "type": "integer",
            "description": "字符串最长长度",
            "format": "int32",
            "nullable": true
          },
          "minLength": {
            "type": "integer",
            "description": "字符串最短长度",
            "format": "int32",
            "nullable": true
          },
          "length": {
            "type": "integer",
            "description": "固定长度",
            "format": "int32",
            "nullable": true
          },
          "min": {
            "type": "integer",
            "description": "数值最小",
            "format": "int32",
            "nullable": true
          },
          "max": {
            "type": "integer",
            "description": "数值最大",
            "format": "int64",
            "nullable": true
          },
          "entityMemberId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "属性的约束更新时请求结构"
      },
      "EntityMemberFilterDto": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "orderBy": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean",
              "nullable": true
            },
            "description": "排序",
            "nullable": true
          },
          "name": {
            "maxLength": 60,
            "type": "string",
            "description": "属性名",
            "nullable": true
          },
          "comment": {
            "maxLength": 300,
            "type": "string",
            "description": "属性注释内容",
            "nullable": true
          },
          "accessModifier": {
            "$ref": "#/components/schemas/AccessModifier"
          },
          "isRequired": {
            "type": "boolean",
            "description": "是否必须",
            "nullable": true
          },
          "isEnum": {
            "type": "boolean",
            "description": "是否为枚举",
            "nullable": true
          },
          "isList": {
            "type": "boolean",
            "description": "是否为列表",
            "nullable": true
          },
          "isObject": {
            "type": "boolean",
            "description": "是否为自定义对象",
            "nullable": true
          },
          "canSet": {
            "type": "boolean",
            "description": "是否可赋值",
            "nullable": true
          },
          "needInit": {
            "type": "boolean",
            "description": "是否要初始化",
            "nullable": true
          },
          "memberType": {
            "$ref": "#/components/schemas/MemberType"
          },
          "constraintId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "objectTypeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "entityModelId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "实体属性查询筛选"
      },
...
      "EntityMemberItemDtoPageList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityMemberItemDto"
            },
            "nullable": true
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "EntityMemberUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 60,
            "type": "string",
            "description": "属性名",
            "nullable": true
          },
          "comment": {
            "maxLength": 300,
            "type": "string",
            "description": "属性注释内容",
            "nullable": true
          },
          "defaultValue": {
            "maxLength": 100,
            "type": "string",
            "description": "默认值，根据类型推断",
            "nullable": true
          },
          "accessModifier": {
            "$ref": "#/components/schemas/AccessModifier"
          },
          "isRequired": {
            "type": "boolean",
            "description": "是否必须",
            "nullable": true
          },
          "isEnum": {
            "type": "boolean",
            "description": "是否为枚举",
            "nullable": true
          },
          "isList": {
            "type": "boolean",
            "description": "是否为列表",
            "nullable": true
          },
          "isObject": {
            "type": "boolean",
            "description": "是否为自定义对象",
            "nullable": true
          },
          "canSet": {
            "type": "boolean",
            "description": "是否可赋值",
            "nullable": true
          },
          "needInit": {
            "type": "boolean",
            "description": "是否要初始化",
            "nullable": true
          },
          "dictionaryKeyType": {
            "$ref": "#/components/schemas/MemberType"
          },
          "dictionaryValueType": {
            "$ref": "#/components/schemas/MemberType"
          },
          "memberType": {
            "$ref": "#/components/schemas/MemberType"
          },
          "objectTypeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "entityModelId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "实体属性更新时请求结构"
      },
...
    }
  }
}